+++ src/main/java/firok/tiths/block/BlockCompressed.java
@@ -1,15 +1,22 @@
 package firok.tiths.block;
 
+import firok.tiths.util.InnerActions;
 import net.minecraft.block.Block;
 import net.minecraft.block.material.MapColor;
 import net.minecraft.block.material.Material;
 import net.minecraft.block.state.IBlockState;
+import net.minecraft.client.util.ITooltipFlag;
+import net.minecraft.item.ItemStack;
 import net.minecraft.util.BlockRenderLayer;
 import net.minecraft.util.EnumFacing;
 import net.minecraft.util.math.BlockPos;
 import net.minecraft.world.IBlockAccess;
+import net.minecraft.world.World;
 import net.minecraftforge.fml.relauncher.Side;
 import net.minecraftforge.fml.relauncher.SideOnly;
+
+import javax.annotation.Nullable;
+import java.util.List;
 
 @SuppressWarnings("all")
 public class BlockCompressed extends Block
@@ -77,6 +84,13 @@ public class BlockCompressed extends Block
 	protected boolean canSilkHarvest()
 	{
 		return true;
+	}
+
+	@SideOnly(Side.CLIENT)
+	@Override
+	public void addInformation(ItemStack stack, @Nullable World player, List<String> tooltip, ITooltipFlag advanced)
+	{
+		InnerActions.addInformation(this, tooltip, advanced);
 	}
 
 }