+++ src/main/java/firok/tiths/block/BlockOre.java
@@ -1,22 +1,26 @@
 package firok.tiths.block;
 
-import firok.tiths.TinkersThings;
+
+import firok.tiths.util.InnerActions;
 import net.minecraft.block.Block;
-import net.minecraft.block.SoundType;
 import net.minecraft.block.material.MapColor;
 import net.minecraft.block.material.Material;
 import net.minecraft.block.state.IBlockState;
+import net.minecraft.client.util.ITooltipFlag;
 import net.minecraft.creativetab.CreativeTabs;
 import net.minecraft.init.Items;
 import net.minecraft.item.Item;
 import net.minecraft.item.ItemStack;
 import net.minecraft.util.NonNullList;
 import net.minecraft.util.math.BlockPos;
-import net.minecraft.util.math.MathHelper;
 import net.minecraft.world.IBlockAccess;
 import net.minecraft.world.World;
+import net.minecraftforge.fml.relauncher.Side;
+import net.minecraftforge.fml.relauncher.SideOnly;
 import slimeknights.tconstruct.library.utils.HarvestLevels;
 
+import javax.annotation.Nullable;
+import java.util.List;
 import java.util.Random;
 
 @SuppressWarnings("all")
@@ -157,7 +161,14 @@ public class BlockOre extends Block
 		return new ItemStack(this);
 	}
 
-//	public int damageDropped(IBlockState p_damageDropped_1_) {
+	@SideOnly(Side.CLIENT)
+	@Override
+	public void addInformation(ItemStack stack, @Nullable World player, List<String> tooltip, ITooltipFlag advanced)
+	{
+		InnerActions.addInformation(this, tooltip, advanced);
+	}
+
+	//	public int damageDropped(IBlockState p_damageDropped_1_) {
 //		return this == Blocks.LAPIS_ORE ? EnumDyeColor.BLUE.getDyeDamage() : 0;
 //	}
 }